Feature/main world /management player object (#45)
* 세션 에 대한 관리 주체 변경 및 해제 수

* Session, Player 생성 주기 수

* Ref 관련 수정 및 코드 정리

* Grid 에  Enqueue 되기전에 패킷 보내는 문제 수

+++ GameServer/ZoneManager/CZoneManager.h
@@ -52,8 +52,8 @@ public:
 public:
 	void SendZone(int Channel, int Zone, CPacket* pPacket, COORDINATE pivot, CPlayer* pPlayer = nullptr);
 public:
-	int m_iLogDelayTime = 1 * 1000;
-	int m_iLogTime = 0;
+	ULONGLONG m_iLogDelayTime = 1 * 1000;
+	ULONGLONG m_iLogTime = 0;
 	void Log();
 
 };
